data-simulator/src/beq_controller/requirements.txt
paho-mqtt
requests
ssl

